reorganize, cleanup, some minor ehancements and changes
Change default gcc-branch to gcc-7-branch so that git pulls will update

Added "make all" "make all install" and point default make and make
install at them.  At the current time gcc is not included in all but is
included in bootstrap-install.

Added make gcc-bootstrap in addition to make gcc as I can't quite
decided whether I want to --enable-bootstrap or --disable-bootstrap,
hence the different make gcc-bootstrap and make gcc.  The former being
more reliable (e.g. when building master from gcc-6) and the latter quicker to build.

Removed gcc-7_2_0-release as an option as it is a tag and not a branch.
This interferred with the default github processing of pulling updates
for the local repository.

Reverted the patch that automatically saved a copy of the gcc-build
after successful install.  Not enough convenience to warrant the hack.